[Tracing SDK][2] Split yaml utils from file_utils.py (#15380)
