+++ mlflow/models/container/__init__.py
@@ -24,8 +24,8 @@ from mlflow.pyfunc import _extract_conda_env, mlserver, scoring_server
 from mlflow.store.artifact.models_artifact_repo import REGISTERED_MODEL_META_FILE_NAME
 from mlflow.utils import env_manager as em
 from mlflow.utils.environment import _PythonEnv
-from mlflow.utils.file_utils import read_yaml
 from mlflow.utils.virtualenv import _get_or_create_virtualenv
+from mlflow.utils.yaml_utils import read_yaml
 from mlflow.version import VERSION as MLFLOW_VERSION
 
 MODEL_PATH = "/opt/ml/model"
